# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My API",
        "description": "API documentation for my Laravel application",
        "contact": {
            "email": "your-email@example.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/transfers": {
            "get": {
                "tags": [
                    "Transfers"
                ],
                "summary": "Get a list of transfers",
                "description": "Retrieve a list of transfers, with optional filtering by driver_id, status, flight_num, and sorting by id, status, pickup, and unload.",
                "operationId": "911e2c0ca9835e3976caf9f8cc13d2d6",
                "parameters": [
                    {
                        "name": "driver_id",
                        "in": "query",
                        "description": "Filter by driver ID",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Filter by status",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "flight_num",
                        "in": "query",
                        "description": "Filter by flight number",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "sort_field",
                        "in": "query",
                        "description": "Field to sort by (e.g., id, status, pickup, unload)",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "sort_direction",
                        "in": "query",
                        "description": "Sort direction (asc or desc)",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of transfers retrieved successfully.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Transfer"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            },
            "post": {
                "tags": [
                    "Transfers"
                ],
                "summary": "Create a new transfer",
                "description": "Create a new transfer record with the provided data.",
                "operationId": "20bcde73df13e86e67ed2eaa878573d8",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Transfer"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Transfer created successfully.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Transfer"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
        "/api/transfers/{id}": {
            "get": {
                "tags": [
                    "Transfers"
                ],
                "summary": "Get a specific transfer",
                "description": "Retrieve a specific transfer by its ID.",
                "operationId": "f20e3907ba605db596ed1d8f476ad520",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Transfer ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Transfer retrieved successfully.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Transfer"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Transfer not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            },
            "put": {
                "tags": [
                    "Transfers"
                ],
                "summary": "Update a specific transfer",
                "description": "Update the specified transfer by its ID.",
                "operationId": "2aea242c6c739e54a9b974773a4b1875",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Transfer ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Transfer"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Transfer updated successfully.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Transfer not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Transfers"
                ],
                "summary": "Delete a specific transfer",
                "description": "Delete the specified transfer by its ID.",
                "operationId": "8da7e65fda5634583caa146863c5da1e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Transfer ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Transfer deleted successfully.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Transfer not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Transfer": {
                "title": "Transfer",
                "description": "Transfer model",
                "required": [
                    "driver_id",
                    "people_qty",
                    "status",
                    "pickup",
                    "unload",
                    "destination"
                ],
                "properties": {
                    "id": {
                        "description": "ID of the transfer",
                        "type": "integer",
                        "example": 1
                    },
                    "driver_id": {
                        "description": "ID of the driver associated with the transfer",
                        "type": "integer",
                        "example": 123
                    },
                    "luggage": {
                        "description": "Details about the luggage",
                        "type": "string",
                        "example": "2 suitcases"
                    },
                    "people_qty": {
                        "description": "Number of people in the transfer",
                        "type": "integer",
                        "example": 3
                    },
                    "flight_num": {
                        "description": "Flight number for the transfer",
                        "type": "string",
                        "example": "AB1234"
                    },
                    "pickup": {
                        "description": "Pickup location",
                        "type": "string",
                        "example": "Airport Terminal 1"
                    },
                    "unload": {
                        "description": "Unload location",
                        "type": "string",
                        "example": "Hotel XYZ"
                    },
                    "gate": {
                        "description": "Gate number at the pickup location",
                        "type": "string",
                        "example": "Gate A"
                    },
                    "destination": {
                        "description": "Transfer destination",
                        "type": "string",
                        "example": "City Center"
                    },
                    "status": {
                        "description": "Current status of the transfer",
                        "type": "string",
                        "example": "completed"
                    },
                    "comment": {
                        "description": "Additional comments for the transfer",
                        "type": "string",
                        "example": "VIP transfer"
                    }
                },
                "type": "object"
            }
        }
    },
...
}